requirements.txt
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.13.0
bitsandbytes>=0.41.0
